Fix Issue #2950 Move CatBoostError into the ai.catboost.common package
