+++ catboost/jvm-packages/catboost4j-prediction/src/test/java/ai/catboost/CatBoostModelTest.java
@@ -1,5 +1,6 @@
 package ai.catboost;
 
+import ai.catboost.common.CatBoostError;
 import junit.framework.TestCase;
 
 import org.junit.jupiter.api.Assertions;
