+++ catboost/jvm-packages/common/src/main/java/ai/catboost/common/CatBoostError.java
@@ -1,4 +1,4 @@
-package ai.catboost;
+package ai.catboost.common;
 
 public class CatBoostError extends Exception {
     public CatBoostError(String message) {
